requirements_dev.txt
pytest==6.0.1
pytest-flask==1.0.0
